+++ src/Pgvector/Npgsql/SparsevecTypeInfoResolverFactory.cs
@@ -1,0 +1,42 @@
+using System;
+using Npgsql.Internal;
+using Npgsql.Internal.Postgres;
+
+namespace Pgvector.Npgsql;
+
+public class SparsevecTypeInfoResolverFactory : PgTypeInfoResolverFactory
+{
+    public override IPgTypeInfoResolver CreateResolver() => new Resolver();
+    public override IPgTypeInfoResolver CreateArrayResolver() => new ArrayResolver();
+
+    class Resolver : IPgTypeInfoResolver
+    {
+        TypeInfoMappingCollection? _mappings;
+        protected TypeInfoMappingCollection Mappings => _mappings ??= AddMappings(new());
+
+        public PgTypeInfo? GetTypeInfo(Type? type, DataTypeName? dataTypeName, PgSerializerOptions options)
+            => Mappings.Find(type, dataTypeName, options);
+
+        static TypeInfoMappingCollection AddMappings(TypeInfoMappingCollection mappings)
+        {
+            mappings.AddType<SparseVector>("sparsevec",
+                static (options, mapping, _) => mapping.CreateInfo(options, new SparsevecConverter()), isDefault: true);
+            return mappings;
+        }
+    }
+
+    sealed class ArrayResolver : Resolver, IPgTypeInfoResolver
+    {
+        TypeInfoMappingCollection? _mappings;
+        new TypeInfoMappingCollection Mappings => _mappings ??= AddMappings(new(base.Mappings));
+
+        public new PgTypeInfo? GetTypeInfo(Type? type, DataTypeName? dataTypeName, PgSerializerOptions options)
+            => Mappings.Find(type, dataTypeName, options);
+
+        static TypeInfoMappingCollection AddMappings(TypeInfoMappingCollection mappings)
+        {
+            mappings.AddArrayType<SparseVector>("sparsevec");
+            return mappings;
+        }
+    }
+}